requirements.txt
aiohttp==3.3.2
aiopg==0.14.0
async-timeout==3.0.0
attrs==18.1.0
certifi==2018.4.16
chardet==3.0.4
idna==2.7
idna-ssl==1.1.0
multidict==4.3.1
psycopg2==2.7.5
PyJWT==1.6.4
requests==2.19.1
SQLAlchemy==1.2.9
urllib3==1.23
yarl==1.2.6
